# docs/openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrailBase",
    "description": "TrailBase APIs",
    "license": {
      "name": "OSL-3.0",
      "identifier": "OSL-3.0"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/auth/v1/logout": {
      "get": {
        "tags": ["auth"],
        "summary": "Logs out the current user and delete **all** pending sessions for that user.",
        "description": "Relies on the client to drop any auth tokens. We delete the session to avoid refresh tokens\nbringing a logged out session back to live.",
        "operationId": "logout_handler",
        "parameters": [
          {
            "name": "redirect_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Auth & refresh tokens."
          }
        }
      },
      "post": {
        "tags": ["auth"],
        "summary": "Logs out the current user and deletes the specific session for the given refresh token.",
        "description": "Relies on the client to drop any auth tokens.",
        "operationId": "post_logout_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Auth & refresh tokens."
          }
        }
      }
    },
...
    "/api/auth/v1/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Registers a new user with email and password.",
        "operationId": "register_user_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful registration."
          }
        }
      }
    },
    "/api/auth/v1/reset_password/request": {
      "post": {
        "tags": ["auth"],
        "summary": "Request a password reset.",
        "operationId": "reset_password_request_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success."
          }
        }
      }
    },
    "/api/auth/v1/reset_password/update/:password_reset_code": {
      "post": {
        "tags": ["auth"],
        "summary": "Endpoint for setting a new password after the user has requested a reset and provided a\nreplacement password.",
        "operationId": "reset_password_update_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success."
          }
        }
      }
    },
...
    "/api/auth/v1/verify_email/trigger": {
      "get": {
        "tags": ["auth"],
        "summary": "Request a new email to verify email address.",
        "operationId": "request_email_verification_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Email verification sent."
          }
        }
      }
    },
...
    "/api/records/v1/{name}/schema": {
      "get": {
        "tags": ["records"],
        "summary": "Retrieve json schema associated with given record api.",
        "operationId": "json_schema_handler",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON schema."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmailVerificationRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
...
      "LogoutRequest": {
        "type": "object",
        "required": ["refresh_token"],
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
      "RegisterUserRequest": {
        "type": "object",
        "required": ["email", "password", "password_repeat"],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "password_repeat": {
            "type": "string"
          }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "ResetPasswordUpdateRequest": {
        "type": "object",
        "required": ["password", "password_repeat"],
        "properties": {
          "password": {
            "type": "string"
          },
          "password_repeat": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}